Experimental implementation of side compactions #979
This change satisfies a very narrow use case of a single tablet on
tablet server  getting lots of bulk imports and having a long running
compaction.  In this case the new bulk imports will not get compacted.
This change a adds a side compaction that compacts new files added to a
tablet while a compaction is running.

For the general case of many tablets getting bulk imports and long
running compactions this change will not help.  The more general
proposal outlined in #564 could handle that situation.