Backported Thrift threaded selector server from 2.0.
To use the threaded selector server, set the following in accumulo-site:

  <property>
      <name>general.rpc.server.type</name>
      <value>THREADED_SELECTOR</value>
  </property>

Accumulo 2.0 uses Thrift version 0.12 which has a bug in the Threaded
selector server (see THRIFT-4847).  Accumulo 1.9 uses Thrift 0.9 which
does not have this bug.  Therefore this patch may work better in
Accumulo 1.9 than Accumulo 2.0 because of the difference in thrift
versions.  Ran some of the ITs that failed in 2.0 and they passed in
1.9.

Two commits were backported from 2.0 for this commit.

Cherry picked ff0478b171b7da8d91d3342f1d972f24dd6e1a62
------------------------------------------------------

Switch to thrift server with multiple selector threads. (#1059)

Before this change Accumulo's used the THsHaServer thrift server. This
server has a single select threads that read and writes frames to
sockets.  The select thread puts the frames on a thread pool for
processing.  On a machine with a lot of cores, this single thread can
become a bottleneck.  This commit switches to use the thrift
TThreadedSelectorServer server which support multiple select threads.

To demonstrate the differernce this change can make, the
RandomCachedLookupsPT performance test in accumulo-testing was run with
and without this change.  The test was run on a machine with 16 cores.
The test loads a table into cache and then does random lookups with
varying numbers of threads.  It measures the average lookup time for
each lookup. Below is a table of running this test.

| Threads | Single Selector | Multi Selector | Speedup |
|---------|-----------------|----------------|---------|
|      1  | 0.14 ms         | 0.14 ms        | 1.00    |
|      4  | 0.18 ms         | 0.19 ms        | 0.95    |
|      8  | 0.24 ms         | 0.24 ms        | 1.00    |
|     16  | 0.44 ms         | 0.44 ms        | 1.00    |
|     32  | 0.85 ms         | 0.81 ms        | 1.05    |
|     64  | 1.74 ms         | 1.40 ms        | 1.24    |
|    128  | 3.51 ms         | 2.44 ms        | 1.44    |
|    256  | 7.21 ms         | 4.52 ms        | 1.60    |

This table shows that with 256 client threads doing lookups that each
lookup took 7.21ms on average with a single selector thread.  With this
change and four selector threads, each loopup took 4.52ms on average
with 256 client threads.  During both test there were 256 threads to
process client request on the tserver.  The selector threads hand frames
off to these 256 threads.

Watching top+jstack while running these test show the single selector
thread will eventually reach 100% on a core.  With 4 selector threads
each used 36% CPU when 256 client threads were doing lookups.

Cherry picked 503f351e0897b640d2256af22f7c22cbe73b7287
------------------------------------------------------

Revert thrift server to custom_hs_ha (#1075)

After the changes in #1059 some of the ITs were failing. The new thrift
server was failing.  This change reverts to the previous thrift server
type as the default while leaving the new type configurable.