WIP adds cache warming
This commits adds a per table setting that will warm files in the cache after
a minor or major compaction.  Performance testing with lots of random scans
interleaved with major compactions showed benefits.  The benefits were seen
while the compaction was running because the files were always in cache and
did not have to compete with the compactions for I/O.

This feature is incomplete because the following situations are not handled.

 * What if warming is turned on but files are not fitting in cache?  This
   could be detected by checking if the first block of a file is still in
   the cache after warming.  But after detecting it, I am not sure what
   action to take.
 * Should warming be conditional on if the tablet was recently queried?