Updates from testing
Added SharedRateLimiterFactory::remove to remove the compaction limiter
from the list when we are done to prevent the background threads from
getting a NPE. Modified the CompactionCoordinator::getCompactionJob
loop to try another server if the first one does not have a job.
Modified Thrift calls so that they did not return null and instead
return an empty object. Modified all RetryableThriftCall<Void> to
RetryableThriftCall<String> so that the object did not retry the
function when there was a successful call.