Get metrics for external compactions working
This change provides external compaction info to existing tserver compaction
metrics code. In theory this should cause tservers to emit metrics for running
and queued external compactions, however it has not been tested yet.

This change cleans up unused ExternalCompactionExecutors and ExternalCompaction
informaiton in the tserver.  This cleanup needed to be done to protect memory
usage and support accurate metrics.

Added strong sanity checks to the sets of files passed between a tablet and
compaction planner. These checks exposed a transient problem that needs further
investigation, but the problem does not prevent compactions.  A tablet is
trying to pass bad information to a planner for a short time and the sanity
prevented the information from being passed.